style: move project directory validation to its own file
